# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "description": "Creates a new user account. Only admins can register users directly. The password must meet complexity requirements. Email must be unique.",
        "operationId": "UsersController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Required role. Allowed: admin | superuser | agent | seller | buyer",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Request successful"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid role header"
          },
          "403": {
            "description": "Forbidden. Required role(s): admin, superuser, agent, seller, buyer"
          },
          "409": {
            "description": "A user with this email already exists"
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "email must be an email",
                        "price must be a positive number"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Register a new user",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "description": "Returns all registered users. Admin-only endpoint.",
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Required role. Allowed: admin | superuser",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Request successful"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserResponseDto"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid role header"
          },
          "403": {
            "description": "Forbidden. Required role(s): admin, superuser"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "List all users",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "description": "Returns a single user by their unique ID. Accessible by admins and agents.",
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": "usr_000001",
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Required role. Allowed: admin | superuser | agent",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Request successful"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid role header"
          },
          "403": {
            "description": "Forbidden. Required role(s): admin, superuser, agent"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2025-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/v1/resource/123"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "description": "Partially updates user fields. Only send fields you want to change. Admin-only.",
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": "usr_000001",
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Required role. Allowed: admin | superuser",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Request successful"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid role header"
          },
          "403": {
            "description": "Forbidden. Required role(s): admin, superuser"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2025-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/v1/resource/123"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email already taken by another user"
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "email must be an email",
                        "price must be a positive number"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Update a user",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "description": "Permanently deletes a user account. Admin-only.",
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": "usr_000001",
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Required role. Allowed: admin | superuser",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Missing or invalid role header"
          },
          "403": {
            "description": "Forbidden. Required role(s): admin, superuser"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2025-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/v1/resource/123"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Delete a user",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Real Estate API",
    "description": "## Real Estate Backend API\n\nThis API powers the real estate platform, supporting property listings, user management, and bookings.\n\n### Authentication & RBAC\nAll protected endpoints require a `role` header indicating the caller's role.\nValid roles: `superuser`, `admin`, `agent`, `seller`, `buyer`\n\n### Response Format\nAll responses follow the envelope: `{ success, statusCode, message, data }`\n\n### Error Format\nAll errors follow: `{ success: false, statusCode, message, errors? }`",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role-header": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "User role for RBAC (superuser | admin | agent | seller | buyer)"
      }
    },
    "schemas": {
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "usr_01J8XKZM"
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "email": {
            "type": "string",
            "example": "jane.doe@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "superuser",
              "admin",
              "agent",
              "seller",
              "buyer"
            ],
            "example": "buyer"
          },
          "phone": {
            "type": "object",
            "example": "+911234567890",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2025-01-15T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "phone",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "Jane Doe",
            "minLength": 2,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "description": "Valid email address — used as login identifier",
            "example": "jane.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "Password: 8–32 characters, must include uppercase, lowercase, digit, and special character",
            "example": "Secure@123",
            "minLength": 8,
            "maxLength": 32
          },
          "role": {
            "type": "string",
            "description": "User role on the platform",
            "enum": [
              "superuser",
              "admin",
              "agent",
              "seller",
              "buyer"
            ],
            "example": "buyer"
          },
          "phone": {
            "type": "string",
            "description": "Contact phone number (E.164 format preferred)",
            "example": "+911234567890"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Updated full name",
            "example": "Jane Smith",
            "minLength": 2,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "description": "Updated email address",
            "example": "jane.smith@example.com"
          },
          "role": {
            "type": "string",
            "description": "Updated role",
            "enum": [
              "superuser",
              "admin",
              "agent",
              "seller",
              "buyer"
            ],
            "example": "agent"
          },
          "phone": {
            "type": "string",
            "description": "Updated phone number",
            "example": "+911234567890"
          }
        }
      },
...
    }
  }
}